requirements/requirements.txt
Django==4.2
djangorestframework~=3.14.0
factory-boy==3.3.0
faker-commerce==1.0.3
mysqlclient==2.1.1
PyMySQL==1.0.2
pytest==8.0.2
pytest-cov==4.1.0
pytest-django==4.8.0
python-decouple==3.8
django-rest-knox==4.2.0
django-cors-headers==4.3.1
